+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -3,3 +3,4 @@ export 'instagram.dart';
 export 'linkedin.dart';
 export 'telegram.dart';
 export 'whatsapp.dart';
+export 'youtube.dart';